+++ satella/os/signals.py
@@ -21,6 +21,7 @@ def hang_until_sig(extra_signals: tp.Optional[tp.Sequence[int]] = None) -> None:
     global end
     extra_signals = extra_signals or ()
 
+    # Set the signal handler
     signal.signal(signal.SIGTERM, __sighandler)
     signal.signal(signal.SIGINT, __sighandler)
     for s in extra_signals:
@@ -28,5 +29,11 @@ def hang_until_sig(extra_signals: tp.Optional[tp.Sequence[int]] = None) -> None:
 
     while not end:
         sleep(0.5, True)
+
+    # Unset the signal handler
+    signal.signal(signal.SIGTERM, signal.SIG_DFL)
+    signal.signal(signal.SIGINT, signal.SIG_DFL)
+    for s in extra_signals:
+        signal.signal(s, signal.SIG_DFL)
 
     end = False  # reset for next use