+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added optional `ujson` support
 * officially supporting Python 3.9
+* fixed a bug in `postcondition`