+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added optional `ujson` support
 * officially supporting Python 3.9
 * fixed a bug in `postcondition`
+* added `write_out_file_if_different`