+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.8.13
 
+* enabled calling `exception_handler` without specifying priority