+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * `call_and_return_stdout` will now raise `TimeoutError` if the task failed to complete
     within given timeout
 * added power support to `AtomicNumber`
+* slight refactor in `AtomicNumber`