+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.8.14
 
+* bugfix for hang_until_sig: it will unset its signal handlers