+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-# v2.8.14
+# v2.8.15
 
-* bugfix for hang_until_sig: it will unset its signal handlers